horizon/django/requirements.txt
# Horizon Django Variant — Requirements
# Parts VI–VII: Production REST API

Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-filter>=24.0
psycopg2-binary>=2.9
pytest>=8.0
pytest-django>=4.8
pytest-cov>=4.0
